refactor: split world::tests to separate files
